working on score
